rag_chatbot/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-chatbot"
version = "0.1.0"
description = "RAG Chatbot for Physical AI Humanoid Robotics Book"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "RAG Chatbot Team"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # FastAPI and related
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    # Vector database
    "qdrant-client>=1.7.0",
    # Database
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",
    # Embeddings
    "cohere>=4.0.0",
    "google-genai>=0.7.0",
    "tiktoken>=0.5.0",
    # LLM integration
    "openai>=1.0.0",
    # Configuration and utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Async and utilities
    "httpx>=0.25.0",
    "aiofiles>=23.0.0",
    "aiohttp>=3.13.2",
    "beautifulsoup4>=4.12.0",
    # System monitoring
    "psutil>=5.9.0",
    "chardet>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.0.0",
]

[tool.setuptools.packages.find]
include = ["databases*", "rag_core*", "shared*", "agents_sdk*", "embedding_pipeline*", "backend*", "frontend*"]
